% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fa2_degree}
\alias{fa2_degree}
\alias{fa2_degree_index}
\title{Degree of each node}
\usage{
fa2_degree(G)

fa2_degree_index(G)
}
\arguments{
\item{G}{Two column matrix with vertex ids.}
}
\value{
A two column matrix. Column one has vertices ids and
column two has its degrees.
}
\description{
Degree of each node

Returns only the degree index
}
\section{Functions}{
\itemize{
\item \code{}: Returns an integer vector with the degree
of each vertex.
}}

