% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fa2_distance}
\alias{fa2_distance}
\title{n-norm distance}
\usage{
fa2_distance(n1, n2, norm = 2)
}
\arguments{
\item{n1}{Numeric vector of coordinates.}

\item{n2}{Numeric vector of coordinates.}

\item{norm}{Order of the norm.}
}
\value{
Scalar. Distance between two vectors (norm).
}
\description{
n-norm distance
}

