% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fa2_force}
\alias{fa2_force}
\title{Estimates the force of a system}
\usage{
fa2_force(pos, G, size, kr = 1, kr2 = 100, ka = 1, nooverlap = FALSE)
}
\arguments{
\item{pos}{A k-column matrix with vertices' positions.}

\item{G}{Two column matrix with vertex ids.}

\item{size}{Numeric vector with vertices' sizes.}

\item{kr}{Repulsion constant}

\item{kr2}{}

\item{ka}{Attraction constant}
}
\description{
Estimates the force of a system
}

