% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fa2_forcei}
\alias{fa2_forcei}
\title{Attracction and repulsion forces between two nodes}
\usage{
fa2_forcei(n1, n2, d1, d2, s1 = 1, s2 = 1, kr = 1, kr2 = 100, ka = 1,
  nooverlap = FALSE)
}
\arguments{
\item{n1}{Position of node 1}

\item{n2}{Position of node 2}

\item{d1}{Degree of node 1}

\item{d2}{Degree of node 2}

\item{s1}{Size of node 1}

\item{s2}{Size of node 2}

\item{kr}{Repulsion constant}

\item{kr2}{}

\item{ka}{Attraction constant}
}
\description{
Attracction and repulsion forces between two nodes
}

