% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fa2_get_center}
\alias{fa2_get_center}
\title{Geometric centre of a graph}
\usage{
fa2_get_center(pos)
}
\arguments{
\item{pos}{K column matrix with vertex positions.}
}
\value{
K length numeric vector pointing to the centre of the
graph.
}
\description{
Geometric centre of a graph
}

